Add a dedicate status method, make URLs shorter.

+++ server/data/export.go
@@ -13,6 +13,6 @@ type DataExport struct {
 	CreatedAt   time.Time      `db:"created_at"`
 	CompletedAt *time.Time     `db:"completed_at"`
 	DownloadURL *string        `db:"download_url"`
-	Progress    float64        `db:"progress"`
+	Progress    float32        `db:"progress"`
 	Args        types.JSONText `db:"args"`
 }
